+++ src/settings.py
@@ -1,6 +1,20 @@
 __doc__ = """
 https://docs.djangoproject.com/en/4.0/topics/logging/
+
+https://stackoverflow.com/a/33186558
+
 """
+import logging
+
+
+class UTCFormatter(logging.Formatter):
+    # this formatter in the file to mitigate circular import
+    converter = logging.time.gmtime
+
+
+#
+# START of LOGGING SETTINGS
+#
 
 DEBUG = "DEBUG"
 INFO = "INFO"
@@ -13,8 +27,7 @@ LOGGING = {
     "disable_existing_loggers": False,
     "formatters": {
         "utc": {
-            # NOTE: you need to include the pythonfile in the classname
-            "()": "logger.UTCFormatter",
+            "()": UTCFormatter,
             "format": "[%(asctime)s.%(msecs)03dZ] %(levelname)s : %(name)s : %(message)s",
             "datefmt": "%Y-%m-%dT%H:%M:%S",
         }