+++ docs/changelog.rst
@@ -2,6 +2,13 @@
 Changelog
 =========
 
+v0.0.2
+======
+
+- add entrypoint
+- removed unused imports
+- mitigated circular import
+
 v0.0.1
 ======